Move output processing into the main session loop
This makes it possible for us orderly shut down a session by terminating
an output channel with nul. Complete input processing is now in a
receiveFrames() goroutine, avoding piping of the input frames through a
channel -- we don't inject the input frames anywhere and it probalby
would not be a good idea.